Changed BuyersAgent to Agent Modified Agent to be more dynamic
Agent now excepts a param to determine
who the agent is representing and that is
used to rename form hook name values and
in turn the data it creates. This page is
still a WIP because it needs validation rules.
Additionally, I updated what fields should based
on the agentType value.