Fixed Client.js bug
- Fixed Client bug that happened when increasing
count for number of buyer/seller. When count was
increased and the client moved to next page and then
navigated back to client page the count was reset.
Additionally, the name field for all values was
migrated to improve json structure for final data
and error fields were changed to reflect that.