Updated AutoComplete
- AutoComplete now shows all available default
suggestions onFocus

Updated Attorney.js

- Attorney.js will now ask if the client has an
attorney already and if they don't it will ask
if they would like a recommendation and introduction.
See TODO in attorney as this feature is a WIP.

Updated ForSaleByOwner

- ForSaleByOwner's data structure has been updated.
I am attempting to ensure that the data can easily be
related to each other and remove duplication in fields
for the rest API that will be consuming the json object.
Additionally, for sale by owner has been broken out into
one additional S.FieldWrapper so the isFSBO question is not
wrapped with other fields.