Removed unneccessary code from InputField
- Inputfield had a useState to handle text in field
however it was not needed and was causing a controlled
vs uncontrolled component error.

Updated RadioSelector

- RadioSelector now has additional field which
can handle a "Other" option. This field allows
the user to type a value into it. It currently
does not work with validation rules however. It
seems this is a pitfall of react-hook-form, but
more research is needed.

Updated Slider

- ref={register} no longer ask if required. This will
be defined in validation rules if it is. This allows the
field to be tracked or not by react-hook-form.

Updated AdditionalInformation

- AdditionalInformation has new inputfields to handle
the question, "Are you with the tracy gagne group" and
render a field or not based on answer.

Updated useCustomFormHook

- useCustomFormHook now also returns watch from
useForm