LenderTest now works as intended
- LenderTest pulls in data from google sheet and
utilizes that data to create autocomplete/dropdown
field that allows a user to select their broker and
broker address. However, if one is selected it will
update the other field(s) to reflect it. This will be
replicated across a few fields.

Updated AutoComplete

- AutoComplete now returns index of suggestion
selected from the original list of suggestions.
This value is returned when using handleOnSelect.