Fixed Address & InputField & useCustomFormHook Bug
- The bug was replicated by typing into the input
field on first load. There was an issue when Address
field was empty it would no recognize the first character.
Upon first review I believed this issue was related to
controlled vs uncontrolled components. But, it was not
the issue was caused by yup validation rules. If the mode
and/or reValidateMode are set to onChange it causes the
issue. It is unknown to me why this is being caused, but it
is likely due to the implementation of yup resolver? and how
it handles controlled inputs. Fixed issue by changing validation
mode and reValidateMode to onBlur.